[kcms/fonts] When adjusting all fonts, keep Small font size smaller
Summary:
By default, the "Small" font is 2 points smaller than the "Normal" one. However when you
adjust the size of all fonts, this 2-pts-smaller relationship is not preserved, so the
Small font is set to the same size of the Normal font, defeating its reason for existence.

This patch preserves the "2px smaller" relationship between the Small and Normal fonts
when adjusting all of them at once.

BUG: 261344
FIXED-IN: 5.19.0

Test Plan: {F8258008}

Reviewers: #plasma, ahmadsamir, bport, cblack

Reviewed By: ahmadsamir, bport, cblack

Subscribers: cblack, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D29156