KCM Plasmasearch: Do not intialize runners in config module
Summary:
The runners are loaded from the plugin info list `Plasma::RunnerManager::listRunnerInfo()`,
because of that there is no need to create a RunnerManager and initialize the plugins.

If you have a lot of additional plugins installed that do a lot of work in the init/reloadConfiguration
method the loading time of the KCM is notable slower.

Additionally the KCMUTILS_VERSION version check has been removed, because
the KF5_MIN_VERSION is 5.69.

Test Plan:
All plugin should show up.
To check this create a debug which prints the length of the `listRunnerInfo` list.
Then you can verify that the length of the list is equal to the count
of entries in the plugin selector. For this you can also use the xdotool utility:
`sleep 5;for i in {1..myNum};  do sleep 0.1;xdotool key Down; done`.

Reviewers: #plasma, davidedmundson, ngraham, meven

Reviewed By: ngraham

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D29319