Fix folderview popup mode list view icon size inheriting icon view icon size
Summary:
The icon size is only selectable in icon view mode.
but selecting a different icon size on icon view mode also changes
the size of the icons in list view mode.

This creates a problem since the list icon size is hardcoded by default to small,
but it seems it is being changed on other places of the code.

BUG: 418269
FIXED-IN: 5.19.0

Test Plan:
I tested this on icon folderview and normal desktop folderview, changing icon sizes in iconview and listview
and restarting plasmashell.

Reviewers: #plasma, #vdg, ngraham

Reviewed By: #vdg, ngraham

Subscribers: ngraham, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D28057