Add FindXorgServer.cmake and use it for touchpad KCM
Summary:
x11 backend of the touchpad KCM uses xserver-properties.h that is part of
xorg-server package. As XORG_INCLUDE_DIRS seems to be empty, so far, it was
relying on Synaptics_INCLUDE_DIRS implicitly and worked only if the header
happened to be in the same directory as Synaptics' which is the case most of
the time.

FindXorgServer.cmake is just a copy of FindXorgLibinput, adapted accordingly.

Test Plan:
Pointed Synaptics_INCLUDE_DIRS somewhere else than /usr/include/xorg and build
succeeded.

Reviewers: #plasma, zzag

Reviewed By: #plasma, zzag

Subscribers: broulik, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D29514