Add a SpinBox (with draggability) to panel thickness adjustment UI
Summary:
We receive many complaints and bug reports about the current UI for adjusting panel
thickness. People don't figure out that they have to drag on the button, and when they do,
they complain that it's too hard to set the value finely enough.

This patch removes the draggable button and replaces it with a more conventional SpinBox,
which in D29534 is given ability to adjust the values by scrolling over it or clicking and
dragging. This yields the following improvements:
- The way you adjust panel thickness is now more obvious
- The existing drag-to-adjust behavior is much nicer and less buggy
- Panel thickness can now be specified numerically
- Fine-tuning panel thickness is now easier using any adjustment method
- We have one fewer custom control to maintain
- Bugs in custom scrolling implementation go away

Touch friendliness is maintained because this SpinBox has an inherently touch-friendly design
with large touchable buttons that can be held down to quickly adjust the value, and touch-and
-drag still works after D29534 is applied.

I wouldn't mind using this SpinBox design elsewhere too; see T9460.

BUG: 418700
BUG: 421169
FIXED-IN: 5.19

Test Plan:
{F8301480}
{F8301388}

Reviewers: abetts, #vdg, #plasma, mart

Reviewed By: #vdg, #plasma, mart

Subscribers: mart, ahiemstra, cblack, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D29535