Drag and drop assigning windows to activities in the default switcher
Summary:
When dragging a window from the taskbar onto the activity switcher
applet, the switcher will pop up showing all the activities. When the
user drops the task onto an activity, it will be moved to said activity.

Alternatively, if the Control key is pressed, the window will be
assigned to the destination activity as well as the current one (copy).

{F8303555}

Reviewers: #plasma, mart, davidedmundson, ngraham

Reviewed By: #plasma, mart

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D29548