Dedicated drop areas for assigning tasks to activities
Summary:
Instead of using a difficult to discover 'move to activity by default,
copy to activity when Control is pressed', show two drop areas when the
mouse hovers an activity item in the switcher.

If the user drops the task onto the left area, the task will be moved
to that activity. If the user drops the task onto the right area, the
task will be assigned to the new activity without removing it from the
current one.

{F8313199}

Reviewers: #plasma, ngraham, mart

Reviewed By: ngraham

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D29679