make bounded() unambiguous
```
std::size_t is the unsigned integer type of the result of the sizeof
operator
```

this can make the bounded call ambiguous as it either sports
an int(int,int) signature or an quint32(quint32,quint32) signature. with
size_t being unsigned and 0 being int it's not necessarily clear to the
compiler which overload to use, so make it clear by using usigned
everywhere.
this specifically fixes arm32 compatibility on ubuntu 20.04