Increase values for touchpad acceleration that can be set from kcm libinput UI
Summary:
Plasma System settings can set only 11 values for touchpad acceleration. The option accept a float number between -1 and 1 but, because of the slider limitation, only 11 values in between multiple of 0,2 con be set.

I think this is a great limitation since some touchpad models need specific values to give the user the proper comfort in handling the pointer. Libinput has already less options than the old synaptic, at least let's give the user the ability to set the acceleration with more precision.

In example, the best value for my touchpad is 0,15, but I can't set it, and even if 0,2 is the nearest, I feel better with 0,15 and want to set it. Unfortunately I can't from the kcm UI and I have to modify //~/.config/touchpadxlibinputrc// config manually. I think this is unacceptable for a DE like Plasma that claims to be highly customizable.

So, in order to increase values that can be set with the slider, I suggest to modify the formula in //touchpad.qml// to pick 41 values in the middle, multiple of 0,05.

Reviewers: #vdg, #plasma, ngraham

Reviewed By: #vdg, ngraham

Subscribers: ndavis, ngraham, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D29365