Free the activity switcher from the shackles of data engines
Most of the activity switcher has been ported away from data engines
a long time ago. The only place where the data engines are used in
the main switcher applet is to call toggleActivityManager dbus method
in PlasmaShell.

This patch introduces the toggleActivityManager member function into
the activities applet backend and ports the main switcher applet to
use it instead of the data engine.