kcmaccess: Allow configuring Orca on Wayland
The Orca screen reader works on Wayland, so allow
to enabled/disable and configure it in the KCM when
in a Wayland session as well.

This removes the restriction that kcmaccess was only
available on X11/XCB, yet disables all tabs except
the screen reader one, since the other features
that could be configured didn't seem to work in a
quick test on Wayland.

BUG: 414546