[GTK Integration] Apply the same theme for GTK2 and GTK3 applications
GTK configuration in Plasma depends on xsettings daemon, that has no
separation between GTK2 and GTK3 theme, therefore separate themes for
GTK2 and GTK3 are impossible, while this daemon is working. The daemon
itself provides the functionality to apply settings to GTK applications
without restarting them.

The functionality to apply different themes to applications, depending on
the framework version isn't necessary, when 99% of the themes are
compatible with both versions and when most of the applications are
using the last GTK version anyway.

Given all that, I think, that a theme selection must be narrowed to one
and only one theme for both versions of the GTK framework.

CCBUG: 423141