[applets/taskmanager] Offer better options when activating grouped tasks
Currently we get many complaints about the behavior of clicking on
grouped tasks, especially for the Icons-Only Task Manager, where the
Present Windows effect is always used when available.

This commit overhauls what happens when you click on a grouped task to
offer three options:
1. Show tooltips (new default setting)
2. Show Present Windows effect
3. Show textual list (AKA group dialog)

The user is now free to choose in the settings window which visualization
best suits them.

This commit does not implement the most heavily-requested option--to
bring forward all of the windows belonging to a grouped task
(https://bugs.kde.org/show_bug.cgi?id=370258)--as it is still blocked by
some technical isues related to stacking order and outstanding UX
uncertainty regarding what should happen when one or more of the windows
in a group happens to beminimized. However, since this commit introduces
a config UI for choosing the desired behavior when activating a grouped
task, there is now a place to put that option once it is implemented.

BUG: 424268
BUG: 390400
CCBUG: 370258
FIXED-IN: 5.20

Depends on https://invent.kde.org/frameworks/plasma-framework/-/merge_requests/39