WIP: [kcms/icons] Turn icon size popup into it's own page
Rather than selecting a a specific item to change and having a
disconnected control that alters the selected item, we can use the space
and present all the icon sizes each with their own control connected
with the label.

We can also reduce the code needed significantly and everything still
works, including key handling.