[applets/taskmanager] Make new "cycle through tasks" option the default
This is ultimately the sanest default as it provides the quickest way to
perform the two most common actions for grouped tasks: switching to the
last-used one, or hunting through all of them to find the one you want.
The "Cycle through task" option allows you to perform both actions with
clicks on the same icon, meaning that you don't have to move your cursor
or your visual focus compared to using any of the other three methods.

CCBUG: 370258