[kcms/fonts] (Re-)Add tooltips using a contextual help button
This commit adds tooltips to font settings for which the label alone isn't
sufficient for laymans to understand their effect. Similar explanations
were previously available in the form of "What's This?"-help but got lost
in commit R119:24b960f92284: "QML port of fonts kcm" perhaps because
tooltip functionality was not readily available at the time. The
tooltips added here are partly based on these older ones.

The tooltips are invoked by either hovering or clicking flat Contextual
Help Buttons. This way the available help is unintrusively visible and
can be easily invoked on touch devices.