[kcms/notifications] Use Kirigami.BasicListItem instead of custom delegate
There's no need since the custom delegate was manufacturing the exact
same look that a Kirigami.BasicListitem can provide with the hidden
separator setting. However this commit does not use that setting, so as
a result the separators now become visible. This makes the list
consistent with other lists in QML-based software.

My sense is that the reason why we hid the lines in the past was
because the lines had a very heavy appearance, but that was improved
recently to look much more lightweight.