Revert "[Folder view] Fix binding loop on width"
This reverts commit 6b1cc98f0959e6118259bf4f932e727aaf6df7a8.

The change may have resolved the binding loop, but it regressed text
display for people without high DPI screens. Apparently the old code was
having the effect of aligning the text to pixel boundaries, while the new
code does not. I will find another way to resolve the binding loop.

BUG: 423511
FIXED-IN: 5.19.5