Integrate standard shortcuts into global shortcuts kcm and rename it to just "Shortcuts"
The standard shortcuts kcm is dropped and instead a new category "Standard Shortcuts" shown in the kcm. The standard shortcuts themselves are divided into some categries like navigation and edit to make finding a particular shortcut easier. The model specific behavior and changing of shortcuts of the model is moved into a base class. Both the new standard shortcuts model and the kglobalaccel model (renamed from ShortcutsModel) inherit from the base and implement their own saving and loading methods. ShortcutsModel now combines both of these model. It is a KConcatenateRowsProxyModel extended to support the tree models that are two levels deep.