[folder view] Smooth out icon size steps
Right now, we have the following icon sizes available:

16 (small)
22 (smallMedium)
32 (medium)
48 (large)
64 (huge)
128 (enormous)

As a result, the differences between the first couple of options are very
small, while the differences between the last two are dramatic. it feels
a bit too abrupt.

This commit removes the smallest size and adds a new 96px size in between 64
and 128 to make the jump less abrupt. Appaopriate 96px places icons were
added in https://invent.kde.org/frameworks/breeze-icons/-/merge_requests/36.

As a result, these are the new icon size options:

22 (smallMedium)
32 (medium)
48 (large)
64 (huge)
96 (large * 2)
128 (enormous)

The default icon size (64) remains the same.

People who were previously using a non-default size that is not the
largest one will experience their icons becoming a bit larger.
(22->32, 32->48, 48->64, 64->96).