[Task Manager] Use cgroup data in fallback match of stream PIDs with applications
Task Manager tries to show an audio icon for the application playing
audio. This is problematic for the case of multi-process applications,
such as a modern browser where multiple PIDs are involved. Various
heuristics have been tried, currently we try to see if the parent PIDs
match anything.

In new Plasma we have cgroup information which puts processes together
when they are launched. This should be a lot more accurate.

This won't work on BSD, but this code is only used in fallback cases
anyway for a non-core feature.

BUG: 417457