[kcms/autostart] Rework entry hiding
The current code extracts the executable name from the Exec line, which is somewhat brittle and unreliable. For this purpose the TryExec line exists. While not all apps have one they can be assumed to be reliable when present.

This check is not relevant for scripts since there we can simply assume they are present, otherwise we wouldn't have found them.

This allows for some refactoring of the code, in particular the removal of the command filed in AutostartEntry.