[kcms/autostart] Don't check whether the executable exists
The model checks whether a desktop file's executable exists and hides the entry if not.

This check is inheritly faulty and falls flat in many cases, e.g. for flatpak'd apps, where the executable is `flatpak`. I don't see a way to reliably verify that an entry is valid.

In @teams/vdg we came to the conclusion that it is better to show the entry that is configured even if it is broken. Otherwise an entry that is accidentally broken may appear nonexistant at all. This reflects the technical state better.