[applets/taskmanager] Change "open new instance" to "open new window"
"Instance" is an esoteric jargon word. Now, technically a new instance of an
app and a new window of the same app are different things, but to most users,
it's a distinction without a difference, and the users who do understand the
difference won't be confused by the use of more user-friendly terminology
anyway.