Add some rudimentary error handling to users kcm
Don't fail silently when an error occurs otherwise the kcm just looks broken.
Show at least an inline message so the user knows that something went wrong.
The comment was wrong. We can distinguish why an error occured. Furthermore,
even if it is a permission denied error we do not know if the user canceled
the action or if it failed for some other reason so also treat that as an
error.
BUG:425036
CCBUG:426932
FIXED-IN:5.20
CCMAIL:uhhadd@gmail.com


(cherry picked from commit 05768142c17600c8d4583f171d08157ecc337390)