Fix sizing of tooltip with media controls without thumbnail
When an application that exposes media controls is pinned to the
taskbar, on a different desktop and the task manager is configured to
"only show tasks from the current desktop", hovering over the entry in
the taskbar will show a cramped and broken tooltip with media controls
overlapping the media title. While this is ugly, it has the additional
effect that all other tooltips end up broken until the tooltip is
reloaded by hovering a grouped entry.

To fix this, the media controls follow the header's maximum width rather
than the width, the two scrolling texts are given a proper implicitWidth
and the media controls are updated to match the header's margins and
spacing.

(cherry picked from commit a2481a944ba3dda006ace5c2c224b59e0da5e816)