Rework the plasmoid configuration dialog
We put a ScrollView/Flickable around the whole content. When we embed a KCM (like we do e.g. in plasma-pa) that comes with its own flickable this creates problems.
This can be seen in the plasma-pa case where it's currently impossible to scroll and the scrollbar placement is wrong.

Instead of having a StackView wrapped by a ScrollView we now have a Loader that can either load a KCM or an applet config (this includes shortcuts and about page).
The applet config is wrapped in a Kirigami.ScrollablePage to allow scrolling if necessary.

BUG: 426998