KCM Notifications use KCModuleData to highlight the module when settings has changed
The particularity is in the per application/service's settings, which are created and loaded dynamically from the model in the KCM, meaning they are not yet available when KCModuleData check for isDefault.

Simple solution introduced here is to allow the KCModuleData derived class to load them from the plasmanotifyrc file since the non default settings are saved here.

Also, show the default indicator in the Applications / Services list and hihglight the "Configure..." button