[panel] Fix dragging panel to resize for top and right panels
Previously, dragging up or right was always required to increase the
value of the spinbox and hance increase the thickness of the panel.
This makes sense for bottom and left panels, but is inverted for top
and right panels.

This commit reverses the stepSize when using a top or right panel, which
has the result that you can now always drag the SpinBox in the direction
of where you want the panel to resize no matter which screen edge the
panel is on.

BUG: 429063
FIXED-IN: 5.20.4