Don't use visible property in procedural code to determine state
It can be invisible not just because something is disabled but because
it's not shown yet. But load() is only run once. Instead replace with a
binding.

Rather than toggling the checked state when invisible, the evaluation is
moved into the sync method.

BUG: 408116


(cherry picked from commit a53b30f3f5d8d1644ce91398ed1457ee9fe525cc)