[plasmoidconfig] Fix sizing for some configurations
We'd like to rely on the config to give us an appropriate
implicitHeight, but some configs do the opposite and assume they fill
the parent. The old code resolved this by making the config at least as
high as its parent. Reintroduce that.

This is arguably all very ugly and should be resolved properly, but that
requires changing the configs themselves, which is infeasible for Plasma
5 with regards to third-party applets

BUG: 429797