Use Instantiator to create drawer actions to improve startup time
Found with qml profiler that the creation of the drawer actions took the most
time. Using an Instantiator is faster and causes the binding to be evaluated
less often. Before the binding inside Kirigami was evaluated 28 times and it was
the thing the most time was spend on, afterwards creating the window is the most
costly operation.
CCBUG:429855


(cherry picked from commit d1462745554739ec86fd0e5311d4f271c7a33f3b)