[kcms/users] Avoid race condition on startup
If accounts service is not yet running the name will not be registered
when we make our first call.

We call into AccountsService from our client code with a blocking call.
We then create a new QDBusAbstractInterface on a given service.

QDBusAbstractInterface::isValid effectively is just a check that the
service is registered. This might not be processed yet at the time we
call this. It isn't a helpful check, if it wasn't up the previous call
to fetch users would have failed. We can just remove it.

BUG: 429314