fix: keyboard variants display incorrectly on upgrade
Earlier, keyboard variants were embedded in layouts in config.
It is displayed incorrectly with new QML Keyboard Layout applet.

Now we need to update the config to split variants into separate
configuration entry.

The code is still compatible with the old format, but should be
refactored in the future.

Config update script sample:
	- Input:
	LayoutList=us,us(intl),ru,gb(dvorak),gb

	- Output:
	LayoutList=us,us,ru,gb,gb
	VariantList=,intl,,dvorak
	# DELETE LayoutList

CCBUG: 390079