[folder view] Fix display on not-skinny vertical panels
Folder view has some logic to switch between its compact and full
representations for vertical panels, but it is faulty, and causes the
switch to happen at approximately 50px width. At this width, the display
is completely unusable as the text is elided away into nothingness.

This commit remedies the situation by imposing a more sane minimum width
based on gridUnit. With the default 10pt font, the minimum thickness is
now 144px, which leaves enough room for the labels to be at least
partially visible. At lower panel widths, the applet displays its compact
representation.

BUG: 406806
FIXED-IN: 5.18.7