[kcms/ksplash] Relocate to Appearance section
The idea behind consolidating all the appearance-related KCMs inside a
top-level "Appearance" category with the Global Themes KCM as the parent
was to communicate to the user which set of settings were controlled by
global themes. It generally works, but was missing a piece: the Splash
Screen KCM. This KCM only controls one thing, and that thing is a part
of global themes. So it logically makes sense to relocate it to the
Appearance section to reinforce its relationship to global themes.