[kcms/keyboard]: Modernise numlock settings
Since the key-held behaviour ported the group box to a formlayout, the numlock settings looked out of place. This modernises them as well so that they use a modern form layout-based design.

Modernisation also includes making behaviour more similar to the held-key behaviour, which notably omits the "don't change" setting which provided a strange UX to users that don't know what's going on behind the scenes of their GUI, as well as not functioning on Wayland, which has nothing analogous to X11's configuration settings that Plasma shouldn't touch.