kickoff: use Kirigami.Avatar directly instead of wrapping in RoundButton
This works around a Kirigami bug in a manner that can be shipped in 5.21, as well as
bringing consistency with other clickable avatars which typically use actions.main
instead of wrapping the Avatar in another component.

As a bonus, we lose a lot of excess code.

CCBUG: 428380


(cherry picked from commit 6a8bf09cfd54abe6dc5789f7cf713e4fec362606)