Fix crash when deleting activity from activity manager
When deleting activity from activity manager, the delete button is destroyed
while handling onClicked signal. This happens because
ActivitySettings.deleteActivity() is calling QMessageBox::question() that
waits for user confirmation.

Change deleteActivity call to use QMetaObject with QueuedConnection as the
connectivity type. This allows the delete button to be destroyed without
having to wait for user confirmation.

BUG: 426926
BUG: 427593