[kcms/touchpad] Cleanup version handling
Instead of defining a TOUCHPAD_KCM_VERSION and adding a header for it
use PROJECT_VERSION directly and let ECM generate a header.

Remove the version from applet and dataengine desktop files. For the
applet it is broken anyway since the file is never configured. Since
apparently nobody noticed until now it makes me question how useful the
version info is, especially given it's the same as the Plasma version.