[applets/kickoff] Make section headers more section headery
The header text uses a PlasmaExtras.Header component, but with the same
size as the text of the list items. We typically use larger text for
section headers to distinguish them. So let's do that.

BUG: 434669
FIXED-IN: 5.21.4


(cherry picked from commit a2dc55f66bbbaf19a7e0b25039a67ee55e137835)