[folder view] Allow disabling click-to-rename when using double-click
a911982efce3efeb510899abb526230b0de12f7f added a "Click on the label to
initiate renaming when using double-click" feature, mimicking the one
that Dolphin has. However Dolphin also has an option to turn it off,
while Folder View thus far does not.

This commit implements a simple setting to turn it off, placed in the
same place in Folder View's config window as the equivalent Dolphin
option. Ths option is not visible and has no effect when using the
default single-click mode.

BUG: 434827
FIXED-IN: 5.22