[kcms/keyboard] tastenbrett: left-align key labels
When the horizontal alignment of a label is not specified, the default
is to follow the natural alignment of the text, e.g. RTL text is aligned
to the right, and LTR text is aligned to the left.

This causes key caps with labels with different directionality to look
quite messed up.