[applets/taskmanager] Un-hardcode tooltip close button icon size
This is a leftover from the time when we had to manually specify icon
sizes to make PC3 ToolButtons look good, and this was accidentally set
to too small an icon size at that time. Today the worked-around bug has
long since been fixed, so we can remove the hardcoding and let the
button use its normal default size.