Set a transientParent for the FolderView ContextMenu
We already had the possibility to specify a visualParent but never did.
The positioning below the parent is not optimal because the Delegates
include the spacing between rows at their bottom which was likely never
discovered because the visualParent was never set. So just set it always
to the cursor position.
CCBUG:417378


(cherry picked from commit 603c0b5a02956765b3dfdab2436a56d1b8122348)