[kcms/notifications] Remove option to hide critical notifications in fullscreen apps
I cannot think of a conceivable use case for this option; it seems to
only exist to let users shoot themselves in the foot, because critical
notifications are critical for a reason and you really should see them
even if you're watching a video or presenting. The most common one is
"low battery" and you really really really do want to see this in those
use cases. So Let's remove this option for safety and sanity.

This commit removes only the UI; the backend code is removed in
https://invent.kde.org/plasma/plasma-workspace/-/merge_requests/832 so
that anyone who for some reason turned this off in the past doesn't lose
the ability to turn it back on and get permanently stuck with missed
critical notifications while they're watching YouTube videos.