Hide tooltip window immediately before opening context menu
Tooltip internally uses TooltipDialog (a QWindow subclass)  which runs a timer
before it hides itself. This is done so it can be reused and to prevent flickering
when switching between Tooltips quickly. However here it causes an issue, when
closing a popup on Wayland, Qt closes all popups that were opened after it. This means
when openening a ContextMenu while a Tooltip is shown, the TooltipDialog will hide
itself *after* the window of the context menu has been created and close it in the
process.
As a workaround we can hide the window manually before creating the context menu.
CCBUG: 417939
FIXED-IN:5.22


(cherry picked from commit d24ab7b5a7fc127f7278be79ef49f9a028358d01)