Delete the Touchpad applet
Ths only thing this applet can do is disable or re-enable your
touchpad. If you have no touchpad, it is pointless. If you have only a
touchpad and no mouse, then the only thing it can do is give you a way
to brick your laptop until you find a mouse you can use to re-enable
the touchpad.

The only conceivable non-catastrophic usage of this applet is to
manually disable a laptop's touchpad when a mouse is plugged in. This is
an extremely niche use case that can already be accomplished by using
the "Toggle Touchpad" keyboard shortcut which is exported by the KCM,
and would be better served by a setting in the touchpad KCM itself to
automatically disable the touchpad when a mouse is plugged in and
re-enable it when unplugged, should such a feature be deemed useful. I
have my doubts, as Libinput has gotten very good at ignoring false input
these days, so I think there is no longer much of a use case for
disabling a touchpad at all.

Overall, this applet is not worth the bugs and degradation of the user
experience that it can cause. It is too dangerous to exist. Let's delete
it.

BUG: 434720
BUG: 434718
BUG: 397210
BUG: 378879
BUG: 430029
BUG: 434717
FIXED-IN: 5.23