[kcms/keyboard] Restore keyboard layout OSD on X11
This has been removed in 9021b3cb4d597413daa8f9960a0b69524fb77715, but I don't quite follow the reasoning.

The commit mentions unifying the codepath with the one used for Wayland which is a sensible goal, but it hasn't happend yet.

There is a known issues with triggering it via the Main shortcut (https://bugs.kde.org/show_bug.cgi?id=408928) but the alternative shortcut works fine.

BUG: 437512
BUG: 437504