[kcms/keyboard] Restore keyboard layout OSD on X11
OSD for alternative shortcut has been removed in 9021b3cb, because that
approach didn't work well with new layout applet.

Here we restore it with original behavior - the OSD is shown when the
shortcut is invoked, but not when used via the applet.

BUG: 435533


(cherry picked from commit 7ff6711523f2ec18e117a39598398e3598b54d18)