KCM Keyboard port to KConfigXT, highlight non default settings
Main goal is to port the KCM to KConfigXT and properly handle default/reset state, and eventually ease the port to QtQuick.
Like other KCMs dealing with shortcuts, they are now automatically applied when selected.
For technical reason, highlight for layouts and tree view keyboard options won't be implemented.

CCBUG: 434075