[kcms/landingpage] Give up on vertical centering and let it scroll
In the past, we made the content view of this KCM center vertically.
However that caused binding loops and made resizing slow. Then later we
made it non-scrollable so that it could be vertically centered, but that
broke scrollability for users with very short screens.

This commit simplifies the situation to stop doing anything fancy at all
and just uses a boring old SimpleKCM. This lets it be scrollable but
does not have vertical centering for the content. However no other KCM
vertically centers its content like this anyway, so arguably there is a
consistency win as well.

BUG: 438171
FIXED-IN: 5.21.2


(cherry picked from commit c561706fed003b1746223a1c25cc4b476c835082)