[containments/desktop] Fix applet overlay icon size with touch interaction
Ths intention of this code was to present larger icons when the config
overlay was shown via a touch event. However this did not work because
the buttons' size was accidentally hardcoded to small, with conflicting
size definitions when only one was needed.

This commit fixes the issue by removing the width and height properties,
using the correct layout properties, and allowing the icon size to
scale ccording to the space available, which fixes the problem for
touch. There are no visual changes with the mouse-based interaction.