[desktoppackage/alternatives] Don't overflow with long text
This code was previously making the list item height equal to the icon
height plus some padding, but the text was vertically unbounded so it
would overflow. This commit fixes that by using the implicit height of
the whole content layout which knows about the height of the labels too.

BUG: 439098
FIXED-IN: 5.22.5


(cherry picked from commit 6749d08aa631e773d5eed58717af49e096a82d47)