[containments/panel] Fix panel applet configuration with touch
Despite this UI seeming very touch-centric, it was actually totally
boken for the touch use due to the use of mouse-specific code, namely
setting the current applet only when the mouse moved and also hiding the
tooltip when the mouse edited the area--neither of which are compatible
with authentic touch events when handled by a mouse area, because the
synthesized mouse position exits the mousearea as soon as the finger
used to touch the screen is retracted.

This is fixed by setting the current applet when clicked/touched as
well, and removing the onExited signal handler to hide the tooltip,
which was unnecessary anyway because the tooltip gets hidden by other
things and this still works.

Now touch interaction with panel applets while in panel edit mode
works as expected.

BUG: 439918
FIXED-IN: 5.22.5


(cherry picked from commit 97add87ba49e751e62f56c03bcaa8ff58c9ad1a4)