[containments/desktop] Stop hardcoding applet handle tooltip position
Doing this caused the tooltip to overlap other buttons when the widget
was rotated 90 degrees horizontally, and therefore made those buttons
inaccessible.

The ToolTip already has code to dodge other UI elements and disappar
when the mouse overs over it; let's let that code work (which it does)
and just remove the hardcoded position overrides here.

BUG: 425636
FIXED-IN: 5.22.5