Wrap file names similar to Dolphin's Icon view mode
This is a follow-up change for new wrapping behavior in KCoreAddons
(and by extension in Dolphin) introduced in frameworks/kcoreaddons!132

https://invent.kde.org/frameworks/kcoreaddons/-/merge_requests/132

Also, editor now relies directly on model data instead of coupling to
label's text. There are many more cases there where direct access is
inappropriate, and should be replaced with properties of model and/or
root item's aliases.