Add more keywords to KCMs
The goal is to increase the number of keywords used to search for
KCMs in System Settings, thus making it easier to find the settings
users want.

Some inaccurate keywords have also been moved or removed, mostnotably
the ones in the mouse KCM that referred to Numpad mouse emulation,
since those settings now live under accessibility.