Make tooltip interactive when there is a player but no window is present
If a media player is pinned to Task Manager, after opening the media
player and switching to another virtual desktop, the media controller
will be still visible in the tooltip but cannot be interacted with.

Move hasPlayer property to ToolTipDelegate.qml, and add
mainItem.hasPlayer to interactive property of ToolTipArea to fix the bug.

BUG: 443425