[kcms/keyboard] Redesign "Add Layout" dialog
Instead of having a three-part language+layout+variant design, it now
has a single list of layouts + search.

The new design is far simpler, and much better reflects how people actually
look for layouts: they don't usually know in which "Layout" category their
language is going to be, and a lot of major languages don't even appear in
the "Language" combobox (like Hindi or Telugu, for example). A lot of users
probably thought their language just wasn't supported because they looked
for it in the "Language" combobox and didn't find it, because they didn't
know it wasn't an exhaustive list of all languages offered in all the
variants. It was just misleading.

With the new design, they just type in the name of their language and done.